apply suggestion and refactor also tests
