+++ cou/apps/factory.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Application class."""
+"""Application factory class."""
 from __future__ import annotations
 
 import logging
