+++ tests/unit/apps/test_auxiliary_subordinate.py
@@ -13,7 +13,12 @@
 #  limitations under the License.
 """Tests of the Auxiliary Subordinate application class."""
 
+import pytest
+
+from cou.apps.auxiliary_subordinate import OvnSubordinateApplication
+from cou.exceptions import ApplicationError
 from cou.steps import UpgradeStep
+from tests.unit.apps.utils import add_steps
 
 
 def test_auxiliary_subordinate(apps):
@@ -46,5 +51,75 @@ def test_auxiliary_subordinate_upgrade_plan_to_victoria(apps, model):
             switch=None,
         ),
     )
+
+    assert upgrade_plan == expected_plan
+
+
+def test_ovn_subordinate(status, model):
+    app = OvnSubordinateApplication(
+        "ovn-chassis",
+        status["ovn_chassis_ussuri_22"],
+        {},
+        model,
+        "ovn-chassis",
+    )
+    assert app.channel == "22.03/stable"
+    assert app.os_origin == ""
+    assert app.apt_source_codename is None
+    assert app.channel_codename == "yoga"
+    assert app.current_os_release == "yoga"
+
+
+def test_ovn_workload_ver_lower_than_22_subordinate(status, model):
+    target = "victoria"
+
+    exp_error_msg_ovn_upgrade = (
+        "OVN versions lower than 22.03 are not supported. It's necessary to upgrade "
+        "OVN to 22.03 before upgrading the cloud. Follow the instructions at: "
+        "https://docs.openstack.org/charm-guide/latest/project/procedures/"
+        "ovn-upgrade-2203.html"
+    )
+
+    app_ovn_chassis = OvnSubordinateApplication(
+        "ovn-chassis",
+        status["ovn_chassis_ussuri_20"],
+        {},
+        model,
+        "ovn-chassis",
+    )
+
+    with pytest.raises(ApplicationError, match=exp_error_msg_ovn_upgrade):
+        app_ovn_chassis.generate_upgrade_plan(target)
+
+
+def test_ovn_subordinate_upgrade_plan(status, model):
+    target = "victoria"
+    app = OvnSubordinateApplication(
+        "ovn-chassis",
+        status["ovn_chassis_ussuri_22"],
+        {},
+        model,
+        "ovn-chassis",
+    )
+
+    upgrade_plan = app.generate_upgrade_plan(target)
+
+    expected_plan = UpgradeStep(
+        description=f"Upgrade plan for '{app.name}' to {target}",
+        parallel=False,
+        function=None,
+    )
+
+    upgrade_steps = [
+        UpgradeStep(
+            description=f"Refresh '{app.name}' to the latest revision of '22.03/stable'",
+            parallel=False,
+            function=model.upgrade_charm,
+            application_name=app.name,
+            channel="22.03/stable",
+            switch=None,
+        ),
+    ]
+    add_steps(expected_plan, upgrade_steps)
 
     assert upgrade_plan == expected_plan
